scheduler/requirements.txt
apscheduler>=3.10
psycopg2-binary>=2.9
httpx>=0.27
openai>=1.0
python-dotenv>=1.0
